rename Fetcher to Backend Soon, it will be handling POST requests too.
